refactor(lib): extract lib/auth_recaptcha.php — reCAPTCHA + login protection (ADR-004, sub of #907)
Moves the reCAPTCHA Enterprise verification and login-protection
widget/CSP/verify functions out of lib.php into lib/auth_recaptcha.php.
ADR-003 conversion applied. Completes the Phase 6 auth four-way split
(auth / auth_password / auth_rate_limit / auth_recaptcha).
