+++ Simple-PHP-IPAM/lib.php
@@ -12,6 +12,7 @@ require_once __DIR__ . '/lib/user_preferences.php';
 require_once __DIR__ . '/lib/auth.php';     // core session + CSRF + login (ADR-004 Task 6.1)
 require_once __DIR__ . '/lib/auth_password.php'; // password policy + reset tokens (ADR-004 Task 6.2)
 require_once __DIR__ . '/lib/auth_rate_limit.php'; // login/IP rate limiting + lockout (ADR-004 Task 6.3)
+require_once __DIR__ . '/lib/auth_recaptcha.php'; // reCAPTCHA + login-form protection (ADR-004 Task 6.4)
 require_once __DIR__ . '/lib/BackupClientInterface.php';
 require_once __DIR__ . '/lib/S3Client.php';
 require_once __DIR__ . '/lib/SftpClient.php';
@@ -3637,296 +3638,10 @@ function parse_sort(array $allowed, string $defaultCol, string $defaultDir = 'as
 }
 
 /* ---------------- Login protection (#124) ---------------- */
-
-/**
- * Verify a reCAPTCHA token against the Enterprise Assessment API.
- * Returns null on pass, error string on fail, null on network error (fail-open).
- *
- * @param array{enabled: bool, project_id: string, api_key: string, expected_action: string, score_threshold: float} $cfg
- */
-function recaptcha_enterprise_verify(string $token, string $siteKey, array $cfg): ?string
-{
-    $projectId      = to_str($cfg['project_id']);
-    $apiKey         = to_str($cfg['api_key']);
-    $expectedAction = to_str($cfg['expected_action']);
-    $threshold      = (float)$cfg['score_threshold'];
-
-    if ($projectId === '' || $apiKey === '') {
-        error_log('reCAPTCHA Enterprise: project_id and api_key must be configured.');
-        return null; // fail open — misconfiguration should not block users
-    }
-
-    $url     = 'https://recaptchaenterprise.googleapis.com/v1/projects/' . rawurlencode($projectId) . '/assessments?key=' . rawurlencode($apiKey);
-    $payload = (string)json_encode(['event' => ['token' => $token, 'expectedAction' => $expectedAction, 'siteKey' => $siteKey]]);
-    $ctx = stream_context_create([
-        'http' => [
-            'method'        => 'POST',
-            'header'        => "Content-Type: application/json\r\nContent-Length: " . strlen($payload) . "\r\n",
-            'content'       => $payload,
-            'timeout'       => 10,
-            'ignore_errors' => true,
-        ],
-        'ssl' => ['verify_peer' => true, 'verify_peer_name' => true],
-    ]);
-
-    try {
-        $raw = @file_get_contents($url, false, $ctx, 0, 1048576);
-        if ($raw === false) {
-            error_log('reCAPTCHA Enterprise: HTTP request failed.');
-            return null;
-        }
-        $resp = json_decode($raw, true);
-        if (!is_array($resp)) {
-            error_log('reCAPTCHA Enterprise: Invalid JSON response.');
-            return null;
-        }
-        /** @var array<string, mixed> $resp */
-        $tokenProps  = is_array($resp['tokenProperties'] ?? null) ? $resp['tokenProperties'] : [];
-        $riskAnalysis = is_array($resp['riskAnalysis'] ?? null)   ? $resp['riskAnalysis']   : [];
-
-        if (!empty($tokenProps['invalid'])) return 'Security check failed. Please try again.';
-        if ($expectedAction !== '' && isset($tokenProps['action']) && to_str($tokenProps['action']) !== $expectedAction) {
-            return 'Security check failed. Please try again.';
-        }
-
-        $score = (isset($riskAnalysis['score']) && is_numeric($riskAnalysis['score'])) ? (float)$riskAnalysis['score'] : 0.0;
-        return $score >= $threshold ? null : 'Security check failed. Please try again.';
-    } catch (Throwable $e) {
-        error_log('reCAPTCHA Enterprise verify error: ' . $e->getMessage());
-        return null;
-    }
-}
-
-/**
- * Resolve the reCAPTCHA v3 expected action name, honouring the legacy
- * top-level $config['recaptcha_action'] key (documented since #289) before
- * falling back to the v2.6.0 registry key recaptcha_enterprise.expected_action.
- * Both the widget render and the Enterprise verify path must go through this
- * helper so the action emitted in the hidden input matches the action checked
- * during verification — otherwise valid Enterprise tokens fail action matching.
- */
-function recaptcha_expected_action_resolved(): string
-{
-    $legacyCfg    = $GLOBALS['config'] ?? null;
-    $legacyAction = is_array($legacyCfg) ? ($legacyCfg['recaptcha_action'] ?? null) : null;
-    $resolved = (is_string($legacyAction) && $legacyAction !== '')
-        ? $legacyAction
-        : to_str(ipam_setting('recaptcha_enterprise.expected_action'));
-    return $resolved !== '' ? $resolved : 'login';
-}
-
-/**
- * Verify the login form protection token/field for the current POST request.
- *
- * Returns null on pass, '' for a silent honeypot rejection (no error shown),
- * or a non-empty error string that should be shown to the user.
- * Fails open on network errors so a broken CAPTCHA provider never blocks login.
- *
- * @param array<string, mixed> $config Stub config (demo_gate) or empty array (login.php); falls back to ipam_setting().
- * @param array<string, mixed> $post
- */
-function login_protection_verify(array $config, array $post): ?string
-{
-    // demo_gate.php passes its own stub; fall back to ipam_setting() for login.php
-    $raw = $config['login_protection'] ?? [];
-    $lp  = is_array($raw) ? $raw : [];
-    $cfg = fn(string $k): mixed => array_key_exists($k, $lp) ? $lp[$k] : ipam_setting("login_protection.{$k}");
-
-    $method = to_str($cfg('method'));
-    if ($method === '' || $method === 'null') return null;
-
-    if ($method === 'honeypot') {
-        return ($post['website'] ?? '') !== '' ? '' : null;
-    }
-
-    if ($method === 'time_check') {
-        $min = max(1, to_int($cfg('min_seconds')));
-        $ts  = to_int($_SESSION['login_form_at'] ?? 0);
-        unset($_SESSION['login_form_at']);
-        if ($ts === 0 || (time() - $ts) < $min) {
-            return 'Form submission was too fast. Please wait a moment and try again.';
-        }
-        return null;
-    }
-
-    $secretKey = to_str($cfg('secret_key'));
-    $siteKey   = to_str($cfg('site_key'));
-
-    if ($method === 'turnstile') {
-        $token = to_str($post['cf-turnstile-response'] ?? '');
-        if ($token === '') return 'Please complete the security check.';
-        try {
-            $resp = oidc_http_post('https://challenges.cloudflare.com/turnstile/v0/siteverify', [
-                'secret'   => $secretKey,
-                'response' => $token,
-                'remoteip' => client_ip(),
-            ]);
-        } catch (Throwable $e) {
-            error_log('Turnstile verify error: ' . $e->getMessage());
-            return null; // fail open
-        }
-        return !empty($resp['success']) ? null : 'Security check failed. Please try again.';
-    }
-
-    if ($method === 'hcaptcha') {
-        $token = to_str($post['h-captcha-response'] ?? '');
-        if ($token === '') return 'Please complete the security check.';
-        try {
-            $resp = oidc_http_post('https://hcaptcha.com/siteverify', [
-                'secret'   => $secretKey,
-                'response' => $token,
-                'remoteip' => client_ip(),
-            ]);
-        } catch (Throwable $e) {
-            error_log('hCaptcha verify error: ' . $e->getMessage());
-            return null; // fail open
-        }
-        return !empty($resp['success']) ? null : 'Security check failed. Please try again.';
-    }
-
-    if ($method === 'recaptcha') {
-        $token = to_str($post['g-recaptcha-response'] ?? '');
-        if ($token === '') return 'Please complete the security check.';
-
-        // Use Enterprise API if configured; fall back to standard reCAPTCHA API
-        if ((bool)ipam_setting('recaptcha_enterprise.enabled')) {
-            $rawThreshold = ipam_setting('recaptcha_enterprise.score_threshold');
-            $enterprise = [
-                'enabled'         => true,
-                'project_id'      => to_str(ipam_setting('recaptcha_enterprise.project_id')),
-                'api_key'         => to_str(ipam_setting('recaptcha_enterprise.api_key')),
-                // Must match the action the widget emits (see
-                // recaptcha_expected_action_resolved for the precedence rules).
-                'expected_action' => recaptcha_expected_action_resolved(),
-                'score_threshold' => is_numeric($rawThreshold) ? (float)$rawThreshold : 0.5,
-            ];
-            return recaptcha_enterprise_verify($token, $siteKey, $enterprise);
-        }
-
-        try {
-            $resp = oidc_http_post('https://www.google.com/recaptcha/api/siteverify', [
-                'secret'   => $secretKey,
-                'response' => $token,
-                'remoteip' => client_ip(),
-            ]);
-        } catch (Throwable $e) {
-            error_log('reCAPTCHA verify error: ' . $e->getMessage());
-            return null; // fail open
-        }
-        return !empty($resp['success']) ? null : 'Security check failed. Please try again.';
-    }
-
-    if ($method === 'friendly_captcha') {
-        $token = to_str($post['frc-captcha-solution'] ?? '');
-        if ($token === '' || $token === '.UNSTARTED') return 'Please complete the security check.';
-        if ($token === '.FETCHING') return null; // in-progress, fail open
-        try {
-            $resp = oidc_http_post('https://api.friendlycaptcha.com/api/v1/siteverify', [
-                'secret'  => $secretKey,
-                'solution'=> $token,
-                'sitekey' => $siteKey,
-            ]);
-        } catch (Throwable $e) {
-            error_log('FriendlyCaptcha verify error: ' . $e->getMessage());
-            return null; // fail open
-        }
-        return !empty($resp['success']) ? null : 'Security check failed. Please try again.';
-    }
-
-    return null; // unknown method — pass through
-}
-
-/**
- * Return the HTML widget snippet to embed in the login/gate form.
- * For time_check, also sets the session timestamp on GET requests.
- *
- * @param array<string, mixed> $config Stub config (demo_gate) or empty array (login.php); falls back to ipam_setting().
- */
-function login_protection_widget_html(array $config): string
-{
-    $raw = $config['login_protection'] ?? [];
-    $lp  = is_array($raw) ? $raw : [];
-    $cfg = fn(string $k): mixed => array_key_exists($k, $lp) ? $lp[$k] : ipam_setting("login_protection.{$k}");
-
-    $method  = to_str($cfg('method'));
-    $siteKey = e(to_str($cfg('site_key')));
-
-    switch ($method) {
-        case 'honeypot':
-            return "<input type='text' name='website' autocomplete='off' tabindex='-1' aria-hidden='true' class='hidden'>";
-        case 'time_check':
-            if ($_SERVER['REQUEST_METHOD'] !== 'POST') {
-                $_SESSION['login_form_at'] = time();
-            }
-            return ''; // no visible widget
-        case 'turnstile':
-            return "<script src='https://challenges.cloudflare.com/turnstile/v0/api.js' async defer></script>"
-                 . "<div class='cf-turnstile' data-sitekey='{$siteKey}'></div>";
-        case 'hcaptcha':
-            return "<script src='https://js.hcaptcha.com/1/api.js' async defer></script>"
-                 . "<div class='h-captcha' data-sitekey='{$siteKey}'></div>";
-        case 'recaptcha':
-            $ver   = to_int($cfg('version'));
-            $isEnt = (bool)ipam_setting('recaptcha_enterprise.enabled');
-            if ($ver === 3) {
-                $scriptSrc    = $isEnt
-                    ? "https://www.google.com/recaptcha/enterprise.js?render={$siteKey}"
-                    : "https://www.google.com/recaptcha/api.js?render={$siteKey}";
-                $entAttr      = $isEnt ? " data-recaptcha-enterprise='1'" : '';
-                $action       = e(recaptcha_expected_action_resolved());
-                $actionAttr   = " data-recaptcha-action='{$action}'";
-                return "<script src='{$scriptSrc}' async defer></script>"
-                     . "<input type='hidden' name='g-recaptcha-response' id='g-recaptcha-response' data-recaptcha-v3-key='{$siteKey}'{$entAttr}{$actionAttr}>";
-            }
-            return "<script src='https://www.google.com/recaptcha/api.js' async defer></script>"
-                 . "<div class='g-recaptcha' data-sitekey='{$siteKey}'></div>";
-        case 'friendly_captcha':
-            return "<script src='https://cdn.jsdelivr.net/npm/friendly-challenge@latest/widget.module.min.js' async defer></script>"
-                 . "<div class='frc-captcha' data-sitekey='{$siteKey}'></div>";
-        default:
-            return '';
-    }
-}
-
-/**
- * Return extra CSP directives needed for the active login protection method.
- * Returns ['script_src' => '...', 'frame_src' => '...'] — either may be empty.
- * Turnstile, hCaptcha, and reCAPTCHA render inside an iframe so frame_src must
- * be explicitly allowed; Friendly Captcha uses Web Components (no iframe needed).
- */
-/**
- * @param array<string, mixed> $config Stub config (demo_gate) or empty array (login.php); falls back to ipam_setting().
- * @return array{script_src: string, style_src: string, frame_src: string}
- */
-function login_protection_extra_csp(array $config): array
-{
-    $raw = $config['login_protection'] ?? [];
-    $lp  = is_array($raw) ? $raw : [];
-    $method = to_str(array_key_exists('method', $lp) ? $lp['method'] : ipam_setting('login_protection.method'));
-    return match ($method) {
-        'turnstile'        => [
-            'script_src' => 'https://challenges.cloudflare.com',
-            'style_src'  => "'unsafe-inline'",
-            'frame_src'  => 'https://challenges.cloudflare.com',
-        ],
-        'hcaptcha'         => [
-            'script_src' => 'https://hcaptcha.com https://assets.hcaptcha.com',
-            'style_src'  => '',
-            'frame_src'  => 'https://newassets.hcaptcha.com',
-        ],
-        'recaptcha'        => [
-            'script_src' => 'https://www.google.com https://www.gstatic.com',
-            'style_src'  => '',
-            'frame_src'  => 'https://www.google.com',
-        ],
-        'friendly_captcha' => [
-            'script_src' => 'https://cdn.jsdelivr.net',
-            'style_src'  => '',
-            'frame_src'  => '',
-        ],
-        default            => ['script_src' => '', 'style_src' => '', 'frame_src' => ''],
-    };
-}
+/* recaptcha_enterprise_verify(), recaptcha_expected_action_resolved(),     */
+/* login_protection_verify(), login_protection_widget_html() and            */
+/* login_protection_extra_csp() moved to lib/auth_recaptcha.php in v3.30.0  */
+/* (ADR-004 Phase 6 Task 6.4, #907).                                        */
 
 /* ---------------- Auto-reserve IPs ---------------- */
